don't delete keypair nor webhook when disconnecting from repo

+++ gitium/gitium.php
@@ -78,7 +78,7 @@ function gitium_deactivation() {
 }
 register_deactivation_hook( __FILE__, 'gitium_deactivation' );
 
-function gitium_uninstall_hook() {
+function gitium_uninstall_hook( $delete_options = true ) {
 	delete_transient( 'gitium_remote_tracking_branch' );
 	delete_transient( 'gitium_remote_disconnected' );
 	delete_transient( 'gitium_uncommited_changes' );
@@ -87,8 +87,10 @@ function gitium_uninstall_hook() {
 	delete_transient( 'gitium_menu_bubble' );
 	delete_transient( 'gitium_is_status_working' );
 
-	delete_option( 'gitium_keypair' );
-	delete_option( 'gitium_webhook_key' );
+	if ( $delete_options ) {
+		delete_option( 'gitium_keypair' );
+		delete_option( 'gitium_webhook_key' );	
+	}
 }
 register_uninstall_hook( __FILE__, 'gitium_uninstall_hook' );
